功能上面的补充

+++ src/ABPExample.Domain/Models/DepartmentMapper.cs
@@ -12,7 +12,7 @@ namespace ABPExample.Domain.Models
 
         public int UserId { get; set; }
 
-        public long DepartmentId { get; set; }
+        public int DepartmentId { get; set; }
 
         public DateTime CreationTime { get; set; }
